+++ server/domain/exceptions.py
@@ -1,10 +1,9 @@
-from abc import ABC, abstractmethod
 from fastapi import HTTPException
 
 class NotFound(HTTPException):
-    def __init__(self, class_obj):
+    def __init__(self, obj):
         self.status_code = 404
-        self.detail = f"{type(class_obj).__name__} Não foi encontrado."
+        self.detail = f"{obj} Não foi encontrado."
 
 class InternalServer(HTTPException):
     def __init__(self):
@@ -12,6 +11,9 @@ class InternalServer(HTTPException):
         self.detail = "Erro no servidor."
 
 class Unauthorized(HTTPException):
-    def __init__(self):
+    def __init__(self, type = "token"):
         self.status_code = 401
-        self.detail = "Token inválido"
+        if type == "token":
+            self.detail = "Token inválido"
+        else:
+            self.detail = "Email ou senha inválido"
